default to schema agnostic font urls
In order to avoid "insecure content" warnings on https hosted sites using semantic ui, `googleProtocol` should default to the schema agnostic "//". This is supported in all modern browsers and avoids the need to switch protocols manually and therefore to rebuild the dist files just in order to deploy on a https page.

btw: great work on semantic ui - I really like it so far!