remove irc channel from readme
The irc channel is being abandoned (lack of use, unanswered questions, etc.).  Most people come, ask a question, and leave after a few minutes.  It should be removed from anywhere else it's mentioned (if anywhere).

If more people joined it'd be a different story, but you need a certain number of active users to make a channel valuable, and that didn't happen.