[Dropdown] Fix issue #3509 and #4041
When using `apiSettings`, the first click on dropdown icon will not
query results. I had change the code to make it focus on the input and
query results in the first click, click on dropdown icon again will
close the menu, but I don't know it shall check for `forceSelection` on
close or not.