Disable refreshing for non active/animated modals
This commit solves #4196
When two or more modals coexist on the same page, the refresh method
of each modal is executed, even if one of them is being shown.
By checking if the module is active or being animated, you prevent
conflicts between shown and hidden modals, like the one described in
the issue.