Use configured lookupFields in search 'get result'
Configuring the lookupFields is a core part of the search module configuration. `get result` was not respecting that and had the default hard-coded. As all it does is pass the value through to the `search object` method (where it is an optional argument), I simply removed the reference.

Failing case: https://jsfiddle.net/15xd2e41/1/
Working case: https://jsfiddle.net/ozcwbcf3/1/